requirements.txt
playwright==1.40.0
asyncio==3.4.3
# Allow a newer psycopg2-binary so pip can pick a prebuilt wheel on Windows
psycopg2-binary>=2.9.9,<3.0.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
fastapi==0.105.0
uvicorn==0.25.0
pydantic==2.5.3
sentry-sdk==1.39.2
prometheus-client==0.19.0
python-telegram-bot==20.7
discord-webhook==1.3.0
beautifulsoup4==4.12.2
lxml==4.9.3